Move BrianQC to its own class... sort of
