+++ psi4/src/psi4/libfock/CMakeLists.txt
@@ -1,4 +1,5 @@
 list(APPEND sources
+  brianqc_int.cc
   CDJK.cc
   CompositeJK.cc
   COSK.cc
